c/Makefile: Adapt to using smflex instead of flex.
I did not have to, but I also renamed lexer1yy.c to lexer1.lex.c.

parser.files: I do not know what, if anything, this is used for
(grep did not find any uses), but to be safe I modified it as well.
Probably it is simply not used and was only checked in by accident.