glr.{h,cc}: Fix failing assertion "referenceCount > 0" in 'StackNode::decRefCt' with 'gcc -O2'.
Previously, with optimization enabled, an assertion would fail in
'StackNode::decRefCt'.  The problem is caused by the fact that the GLR
algorithm can create a cyclic parse node graph, which is leaked because
there is no cycle collection system.  In turn, depending on the order of
objects in the node pool (hence the dependence on compiler
optimization), a node with a self-cycle could decrement the refct of
another object it pointed at twice, thus failing the assertion.

The fix here is, in ~StackNode, to remove all sibling links (including
recursively) before actually destroying the object.  However, that does
nothing to reclaim parse node cycles--it merely stops them from messing
up the reference counts.  Reclaiming cycles while the parser runs will
have to wait for a future fix.  (They get cleaned up when parsing
finishes by the object pool.)

Reported by user mingodad, but I've seen it before.