Adapt to 'string' being 'std::string'.
The changes related to EmitCode and "failed()" are actually fixing a
long-standing bug.  When using OldSmbaseString, there was an implicit
conversion to bool (via the conversion to char*), and that was what
the boolean tests of EmitCode were doing, which is useless.  The
intent was to check for an error opening the file, which is now what
it does.